requirements.txt
# requirements.txt
fastapi>=0.104.0
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1
motor==3.7.0
python-dotenv>=1.0.0
httpx>=0.25.0
uuid>=1.30
python-multipart>=0.0.6
langchain==0.3.20
openai==1.66.3
PyPDF2==3.0.1
ffmpeg-python==0.2.0
uuid>=1.30
google-generativeai==0.8.5
google-genai==1.11.0
youtube_transcript_api==1.0.3
# LLM Providers - uncomment as needed
# anthropic>=0.5.0
# llama-cpp-python>=0.2.11
PyMuPDF==1.25.5
PyDrive==1.3.1
selenium==4.32.0
urllib3==2.3.0
youtube-transcript-api==1.0.3
gdown==5.2.0
firecrawl-py==1.14.1
beautifulsoup4==4.13.4


